requirements.txt
django==4.2.4
argon2-cffi==21.3.0
python-decouple==3.8
dj-database-url==2.0.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
markdown==3.4.4
django-filter==23.2
django-cors-headers==4.2.0
timeflake==0.4.2
django-behaviors==0.5.1
django-simple-history==3.3.0
whitenoise[brotli]==6.5.0
psycopg2-binary==2.9.7
